+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## TBD
 
+### Enhancements
+
+* Add `Configuration.MaxStringValueLength` config option. [#668](https://github.com/bugsnag/bugsnag-unity/pull/668)
+
 ### Dependency updates
 
 * Update bugsnag-cocoa from v6.24.0 to [v6.25.0](https://github.com/bugsnag/bugsnag-cocoa/blob/master/CHANGELOG.md#6240-2022-10-05)